corrected pre download clean

+++ script/download-file/customize.py
@@ -13,7 +13,7 @@ def preprocess(i):
     quiet = (env.get('CM_QUIET', False) == 'yes')
 
     tool = env.get('CM_DOWNLOAD_TOOL', '')
-    pre_clean = env.get('CM_PRE_DOWNLOAD_CLEAN', False)
+    pre_clean = env.get('CM_PRE_DOWNLOAD_CLEAN', True)
 
     #    xsep = '^&^&' if windows else '&&'
     xsep = '&&'
@@ -208,7 +208,8 @@ def preprocess(i):
         for x in ['CM_DOWNLOAD_CMD', 'CM_DOWNLOAD_CHECKSUM_CMD']:
             env[x+'_USED']='YES' if env.get(x,'')!='' else 'NO'
     else:
-        env['CM_PRE_DOWNLOAD_CLEAN_CMD'] = "rm -f {}".format(env['CM_DOWNLOAD_FILENAME'])
+        if pre_clean:
+            env['CM_PRE_DOWNLOAD_CLEAN_CMD'] = "rm -f {}".format(env['CM_DOWNLOAD_FILENAME'])
 
     return {'return':0}
